requirements.txt
# ERP/MES Simulator Dependencies
# Python 3.11+

# ============================================
# API Server (FastAPI)
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database - Async
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# WebSocket
websockets>=12.0

# ============================================
# Data Generation
# ============================================
faker>=24.0.0
numpy>=1.26.0
pandas>=2.2.0

# ============================================
# Configuration
# ============================================
pyyaml>=6.0.1
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# ============================================
# Progress & Logging
# ============================================
tqdm>=4.66.0
rich>=13.7.0
structlog>=24.1.0

# ============================================
# CLI
# ============================================
click>=8.1.7
typer>=0.9.0

# ============================================
# Testing
# ============================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
